Hoist the replace, insert and patch methods into the worker class
This change removes the indirection of having replace, insert and patch
methods in the executor and the Worker. Now the methods are directly in
the worker and are tested at that level as well.

Change-type: patch
